Decompile NO3 EntityFireWargWaveAttack
